feat: LocalTime에 대한 확장 함수 생성, spinner class naming 변경
