+++ app/src/main/java/woowacourse/movie/BookedDateSpinnerAdapter.kt
@@ -1,0 +1,42 @@
+package woowacourse.movie
+
+import android.view.View
+import android.view.ViewGroup
+import android.widget.BaseAdapter
+import android.widget.TextView
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
+
+class BookedDateSpinnerAdapter(val value: List<LocalDate>) : BaseAdapter() {
+    override fun getCount(): Int {
+        return value.size
+    }
+
+    override fun getItem(position: Int): LocalDate {
+        return value[position]
+    }
+
+    override fun getItemId(position: Int): Long {
+        return position.toLong()
+    }
+
+    override fun getView(
+        position: Int,
+        convertView: View?,
+        parent: ViewGroup
+    ): View {
+        val view: TextView
+
+        if (convertView == null) {
+            view = TextView(parent.context)
+            view.text = view.context.getString(
+                R.string.movie_running_dateTime,
+                getItem(position).toDotFormat()
+            )
+        } else {
+            view = convertView as TextView
+        }
+
+        return view
+    }
+}
